requirements.txt
google-cloud-vision
opencv-python
numpy
pytesseract
Pillow
python-dotenv
